requirements.txt
# Core dependencies
rich>=14.0.0
flask>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
rapidfuzz>=3.0.0
requests>=2.31.0
markdown>=3.5.1

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
